Document and Restructure Frontend
