+++ apps/frontend/src/components/WordDialog.tsx
@@ -1,33 +1,50 @@
+/**
+ * @fileoverview This component displays a dialog with information about a word.
+ * It includes a GPT-powered explanation, dictionary definitions, and the ability to save a clip of the word being used.
+ */
+
 import React, { useEffect, useState, useRef } from "react";
 import { motion, useDragControls } from "framer-motion";
 import ReactMarkdown from "react-markdown";
 import remarkGfm from "remark-gfm";
 import remarkBreaks from "remark-breaks";
 import { Copy, Check, Bookmark } from "lucide-react";
-import { apiFetch, ApiError } from "../utils/api";
+import { apiFetch } from "../services/api";
 import { toast } from "react-hot-toast";
-import { toastApiError } from "../utils/error_toaster";
+import { apiWordQuery } from "../services/dictApi";
+import { toastApiError } from "../utils/apiErrorToaster";
 import {
-    apiWordQuery,
+    GptTemplate,
+    SaveClipResponse,
+    WordDialogProps,
     DictLookup,
     JMEntry,
     JMNEntry,
     KanjiInfo,
-} from "../utils/dict_api";
-import { GptTemplate, SaveClipResponse } from "../types";
+    ApiError,
+} from "../types/types";
+import { createAndSaveClip } from "../utils/clipCreator";
 
-interface Props {
-    sentence: string;
-    word: string;
-    onClose: () => void;
-    cueStart: number;
-    cueEnd: number;
-    videoFile: File | null;
-    videoUrl?: string;
-}
+import { createRecordingPromise } from "../utils/mediaRecorder";
+import {
+    JmdictEntryDisplay,
+    JmnedictEntryDisplay,
+    KanjiInfoDisplay,
+    ExampleDisplay,
+} from "./DictDisplays";
 
 const gptCache = new Map<string, any>();
 
+/**
+ * The WordDialog component.
+ *
+ * This component is responsible for the following:
+ * - Displaying information about a word, including a GPT-powered explanation and dictionary definitions.
+ * - Allowing the user to save a clip of the word being used in the video.
+ *
+ * @param {WordDialogProps} props The props for the component.
+ * @returns {JSX.Element} The WordDialog component.
+ */
 export default function WordDialog({
     sentence,
     word,
@@ -36,7 +53,7 @@ export default function WordDialog({
     cueEnd,
     videoFile,
     videoUrl,
-}: Props) {
+}: WordDialogProps) {
     const key = `${sentence}__${word}`;
     const [data, setData] = useState<any | null>(gptCache.get(key) ?? null);
     const [tab, setTab] = useState<"gpt" | "dict">("dict");
@@ -216,361 +233,58 @@ ${dictData.meanings.join(", ")}`;
         }
     };
 
-    const createRecordingPromise = (
-        stream: MediaStream,
-        duration: number
-    ): Promise<File> => {
-        return new Promise<File>((resolve, reject) => {
-            let recorder: MediaRecorder;
-            const chunks: BlobPart[] = [];
-            if (duration <= 0) {
-                return reject(
-                    new Error("Recording duration must be positive.")
-                );
-            }
-            try {
-                if (
-                    !stream ||
-                    !stream.active ||
-                    stream.getTracks().length === 0
-                ) {
-                    return reject(
-                        new Error(
-                            "Stream is invalid or has no tracks at MediaRecorder creation."
-                        )
-                    );
-                }
-                const options = { mimeType: "video/webm;codecs=vp8,opus" };
-                try {
-                    recorder = new MediaRecorder(stream, options);
-                } catch (e) {
-                    console.warn(
-                        `Recorder init failed with ${options.mimeType}, trying default:`,
-                        e
-                    );
-                    recorder = new MediaRecorder(stream);
-                }
-                recorder.ondataavailable = (event) => {
-                    if (event.data && event.data.size > 0)
-                        chunks.push(event.data);
-                };
-                recorder.onstop = () => {
-                    if (chunks.length === 0)
-                        return reject(new Error("No video data recorded."));
-                    const clipMimeType = recorder.mimeType || "video/webm";
-                    const clipBlob = new Blob(chunks, { type: clipMimeType });
-                    if (clipBlob.size === 0)
-                        return reject(new Error("Recorded clip is empty."));
-                    const ext =
-                        clipMimeType.split(";")[0].split("/")[1] || "webm";
-                    resolve(
-                        new File([clipBlob], `clip.${ext}`, {
-                            type: clipMimeType,
-                        })
-                    );
-                };
-                recorder.onerror = (event) =>
-                    reject(new Error("MediaRecorder error"));
-                recorder.start();
-                setTimeout(() => {
-                    if (recorder?.state === "recording") recorder.stop();
-                }, duration);
-            } catch (setupError) {
-                console.error(
-                    "Error in createRecordingPromise setup:",
-                    setupError
-                );
-                reject(setupError);
-            }
-        });
-    };
-
     const handleSave = async () => {
         if (!canSaveClip) {
             toast.error("No video source available");
             return;
         }
-        const videoElement = document.getElementById(
-            "mirumoji-player"
-        ) as HTMLVideoElement;
-        if (!videoElement) {
-            toast.error("Video player not found.");
-            setSaving(false);
-            return;
-        }
-        let adjustedCueEnd = cueEnd + 1.0;
-        if (cueStart >= adjustedCueEnd) {
-            toast.error("Invalid clip duration.");
-            setSaving(false);
-            return;
-        }
-        // Try to get Video Duration so that clip extension doesn't exceed it
-        const videoDuration = videoElement.duration;
-
-        // Only clamp if videoDuration is a valid, finite number
-        if (
-            typeof videoDuration === "number" &&
-            !isNaN(videoDuration) &&
-            isFinite(videoDuration)
-        ) {
-            if (cueStart >= videoDuration) {
-                toast.error("Clip start time is at or after video end.");
-                setSaving(false);
-                return;
-            }
-            if (adjustedCueEnd > videoDuration) {
-                console.warn(
-                    `Adjusted cue end (${adjustedCueEnd}s) exceeds video duration (${videoDuration}s). Clamping to video duration.`
-                );
-                adjustedCueEnd = videoDuration;
-            }
-        } else {
-            console.warn(
-                "Video duration is not available or not finite (e.g., live stream). Using original cueEnd"
-            );
-            adjustedCueEnd = cueEnd;
-        }
         setSaving(true);
-        let gptData = data;
-
-        if (!gptData) {
-            toast.loading("Fetching GPT", {
-                id: "gptSaveToast",
-            });
-            try {
-                gptData = await fetchGptData();
-                toast.dismiss("gptSaveToast");
-            } catch (error) {
-                setSaving(false);
-                toast.dismiss("gptSaveToast");
-                return;
-            }
-        }
-
-        if (!gptData) {
-            setSaving(false);
-            toast.error("Could not fetch GPT. Please try again.");
-            return;
-        }
-
-        const originalPlayerTime = videoElement.currentTime;
-        const wasPlayerPaused = videoElement.paused;
-        const originalMutedState = videoElement.muted;
-        let streamToCleanup: MediaStream | null = null;
 
         try {
-            const recordingDurationMs = (adjustedCueEnd - cueStart) * 1000;
-            if (recordingDurationMs <= 0) {
-                throw new Error("Clip duration must be positive.");
+            let gptData = data;
+            if (!gptData) {
+                toast.loading("Fetching GPT data...", { id: "gpt-fetch" });
+                gptData = await fetchGptData();
+                toast.dismiss("gpt-fetch");
             }
 
-            videoElement.currentTime = cueStart;
-            videoElement.muted = true;
-            await videoElement
-                .play()
-                .catch((e) => console.warn("Playback warning during save:", e));
-            await new Promise((r) => setTimeout(r, 150));
-
-            const capturedStream = videoElement.captureStream();
-            streamToCleanup = capturedStream;
-
-            if (
-                !capturedStream ||
-                !capturedStream.active ||
-                capturedStream.getTracks().length === 0
-            ) {
-                throw new Error(
-                    "Failed to capture a valid video stream for clipping."
-                );
+            if (!gptData) {
+                toast.error("Could not fetch GPT data. Please try again.");
+                setSaving(false);
+                return;
             }
 
-            const clipFile = await createRecordingPromise(
-                capturedStream,
-                recordingDurationMs
+            const onProgress = (
+                message: string,
+                type: "success" | "error" | "loading"
+            ) => {
+                const toastId = "clip-save-toast";
+                switch (type) {
+                    case "loading":
+                        toast.loading(message, { id: toastId });
+                        break;
+                    case "success":
+                        toast.success(message, { id: toastId });
+                        break;
+                    case "error":
+                        toast.error(message, { id: toastId });
+                        break;
+                }
+            };
+
+            await createAndSaveClip(
+                "mirumoji-player", // The ID of the main video player
+                cueStart,
+                cueEnd,
+                gptData,
+                onProgress
             );
-
-            videoElement.currentTime = originalPlayerTime;
-            if (wasPlayerPaused && !videoElement.paused) videoElement.pause();
-            else if (!wasPlayerPaused && videoElement.paused)
-                await videoElement
-                    .play()
-                    .catch((e) => console.warn("Restore play failed:", e));
-            videoElement.muted = originalMutedState;
-
-            if (streamToCleanup) {
-                streamToCleanup.getTracks().forEach((track) => track.stop());
-                streamToCleanup = null;
-            }
-
-            const formData = new FormData();
-            formData.append("clip_start_time", cueStart.toString());
-            formData.append("clip_end_time", adjustedCueEnd.toString());
-            formData.append("gpt_breakdown_response", JSON.stringify(gptData));
-            formData.append("video_clip", clipFile, clipFile.name);
-
-            const toastId = "saveClipToast";
-            toast.loading("Saving...", { id: toastId });
-
-            const response = await apiFetch<SaveClipResponse>(
-                "/profiles/clips/save",
-                { method: "POST", body: formData }
-            );
-
-            if (response.success) {
-                toast.success(response.message || "Clip Saved!", {
-                    id: toastId,
-                });
-            } else {
-                throw new Error(response.message || "Failed to save clip.");
-            }
-        } catch (error: any) {
-            console.error("Error during clip saving process:", error);
-            toast.error(
-                error.message ||
-                    "An unexpected error occurred while saving the clip."
-            );
-            if (videoElement) {
-                videoElement.currentTime = originalPlayerTime;
-                if (wasPlayerPaused && !videoElement.paused)
-                    videoElement.pause();
-                else if (!wasPlayerPaused && videoElement.paused)
-                    await videoElement
-                        .play()
-                        .catch((e) =>
-                            console.warn("Error restore play failed:", e)
-                        );
-                videoElement.muted = originalMutedState;
-            }
+        } catch (error) {
+            console.error("Failed to save clip from WordDialog:", error);
         } finally {
-            if (streamToCleanup) {
-                streamToCleanup.getTracks().forEach((track) => track.stop());
-            }
             setSaving(false);
         }
     };
-
-    const JmdictEntryDisplay = ({
-        entry,
-        isLast,
-    }: {
-        entry: JMEntry;
-        isLast: boolean;
-    }) => (
-        <div className={`py-2 ${!isLast ? "border-b border-neutral-700" : ""}`}>
-            <div className="flex items-center">
-                <h3 className="text-lg font-bold mr-2">
-                    {entry.kanji.join("、")}
-                </h3>
-                <p className="text-md text-neutral-300">
-                    {entry.kana.join("、")}
-                </p>
-            </div>
-            {entry.senses.map((sense, i) => (
-                <div key={i} className="ml-4 mt-1">
-                    <p className="text-neutral-400 text-sm">({sense.pos})</p>
-                    <p>
-                        <span className="text-neutral-400">{i + 1}.</span>{" "}
-                        {sense.gloss}
-                    </p>
-                </div>
-            ))}
-        </div>
-    );
-
-    const ExampleDisplay = ({
-        example,
-        key,
-        isLast,
-    }: {
-        example: string;
-        key: number;
-        isLast: boolean;
-    }) => (
-        <div className={`py-2 ${!isLast ? "border-b border-neutral-700" : ""}`}>
-            <div className="flex items-center">
-                <div key={key} className="ml-4 mt-1">
-                    <p className="text-neutral-400 text-lg">({example})</p>
-                </div>
-            </div>
-        </div>
-    );
-
-    const JmnedictEntryDisplay = ({
-        entry,
-        isLast,
-    }: {
-        entry: JMNEntry;
-        isLast: boolean;
-    }) => (
-        <div className={`py-2 ${!isLast ? "border-b border-neutral-700" : ""}`}>
-            <div className="flex items-center">
-                <h3 className="text-lg font-bold mr-2">
-                    {entry.kanji.join("、")}
-                </h3>
-                <p className="text-md text-neutral-300">
-                    {entry.kana.join("、")}
-                </p>
-            </div>
-            <p className="text-neutral-400 text-sm">
-                ({entry.translation_type})
-            </p>
-            <p>{entry.gloss.join("; ")}</p>
-        </div>
-    );
-
-    const KanjiInfoDisplay = ({
-        kanjiInfo,
-        isLast,
-    }: {
-        kanjiInfo: KanjiInfo;
-        isLast: boolean;
-    }) => (
-        <div className={`py-2 ${!isLast ? "border-b border-neutral-700" : ""}`}>
-            <h3 className="text-xl font-bold">{kanjiInfo.literal}</h3>
-            <div className="grid grid-cols-2 gap-2 text-sm mt-1">
-                <p>
-                    <span className="font-semibold text-neutral-400">
-                        Strokes:
-                    </span>{" "}
-                    {kanjiInfo.stroke_count}
-                </p>
-                <p>
-                    <span className="font-semibold text-neutral-400">
-                        Grade:
-                    </span>{" "}
-                    {kanjiInfo.grade || "N/A"}
-                </p>
-                <p>
-                    <span className="font-semibold text-neutral-400">
-                        JLPT:
-                    </span>{" "}
-                    {kanjiInfo.jlpt_tanos || kanjiInfo.jlpt_kanjidic || "N/A"}
-                </p>
-                <div className="col-span-2">
-                    <p>
-                        <span className="font-semibold text-neutral-400">
-                            On'yomi:
-                        </span>{" "}
-                        {kanjiInfo.onyomi.join("、")}
-                    </p>
-                    <p>
-                        <span className="font-semibold text-neutral-400">
-                            Kun'yomi:
-                        </span>{" "}
-                        {kanjiInfo.kunyomi.join("、")}
-                    </p>
-                </div>
-                <div className="col-span-2">
-                    <p>
-                        <span className="font-semibold text-neutral-400">
-                            Meanings:
-                        </span>{" "}
-                        {kanjiInfo.meanings.join(", ")}
-                    </p>
-                </div>
-            </div>
-        </div>
-    );
 
     return (
         <div className="fixed inset-0 bg-black/60 flex items-center justify-center z-50 pointer-events-none p-4">
